pyproject.toml
[project]
name = "cycling-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Sandro", email = "sandro.rodrigues@dipcode.com" }
]
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.95.0",
    "click>=8.3.2",
    "dateparser>=1.4.0",
    "deepagents>=0.5.2",
    "facebook-business>=25.0.1",
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.0",
    "langgraph>=1.1.6",
    "pillow>=12.2.0",
    "polyline>=2.0.4",
    "py-staticmaps>=0.4.0",
    "pydantic>=2.13.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.0.0",
    "python-telegram-bot[job-queue]>=22.7",
    "pyyaml>=6.0.3",
    "rich>=15.0.0",
    "sqlalchemy>=2.0.49",
    "stravalib>=2.4",
    "structlog>=25.5.0",
    "tzdata>=2026.1",
]

[project.scripts]
cycling-agent = "cycling_agent.cli:main"

[build-system]
requires = ["uv_build>=0.9.1,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "freezegun>=1.5.5",
    "pre-commit>=4.5.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "respx>=0.23.1",
    "ruff>=0.15.10",
    "ty>=0.0.30",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "RUF", "ASYNC"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers --cov=cycling_agent --cov-report=term-missing"

[tool.ty.src]
root = "src"
